bin/modules/pyproject.toml
[tool.poetry]
name = "bao-base-bin-modules"
version = "0.1.0"
description = "Fallback dependencies for scripts"
authors = ["rootminus0x1"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9"
pandas = "^2.2.3"
tabulate = "^0.9.0"
toml = "^0.10.2"
packaging = "^24.2"
safe-pysha3 = "^1.0.4"
